+++ day25/src/app/boards/[boardId]/page.tsx
@@ -7,3 +7,4 @@ export default function BoardsDetailPage() {
 }
 
 //리팩토링 = 큰 파일을 분리해서 역할을 명확하게 나누는 것
+
