+++ day25/src/components/boards-detail/hook.ts
@@ -1,0 +1,17 @@
+
+import { useQuery } from "@apollo/client";
+import { FETCH_BOARD } from "./queries";
+import { IBoardDetail } from "./types";
+
+export const useBoardDetail = (boardId: string) => {
+  const { data, loading, error } = useQuery<{ fetchBoard: IBoardDetail }>(
+    FETCH_BOARD,
+    { variables: { boardId } }
+  );
+
+  return {
+    board: data?.fetchBoard,
+    loading,
+    error,
+  };
+};
